refactor root & api routes
