+++ src/app/layout.tsx
@@ -1,0 +1,29 @@
+import "@@/styles/globals.css";
+import type { Metadata } from "next";
+import { Inter } from "next/font/google";
+
+const inter = Inter({ subsets: ['latin'] });
+
+export const metadata: Metadata = {
+    title: "Nextjs Discord Bot Boilerplate",
+    description: ""
+}
+
+export default function RootLayout({
+    children,
+}: { children: React.ReactNode }) {
+
+    return (
+        <html lang="en">
+            <head>
+                <title>Next Js Discord Bot Boilerplate | Mark Matthew Vergara</title>
+                <meta name="description" content="Generated by create next app" />
+                <meta name="viewport" content="width=device-width, initial-scale=1" />
+                <link rel="icon" href="/favicon.ico" />
+            </head>
+            <body className={`${inter.className}`}>
+                {children}
+            </body>
+        </html>
+    )
+}
