--- src/pages/_app.tsx
@@ -1,6 +1,0 @@
-import type { AppProps } from "next/app";
-import "../styles/globals.css";
-
-export default function App({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />;
-}
