--- src/pages/index.tsx
@@ -1,66 +1,0 @@
-import { FormEvent, useState } from "react";
-import { Inter } from "next/font/google";
-import Head from "next/head";
-import axios from "axios";
-
-const inter = Inter({ subsets: ["latin"] });
-
-export default function Home() {
-  const [registerCommandsKey, setRegisterCommandsKey] = useState<string>("");
-  const [status, setStatus] = useState<string>("");
-
-  const handleRegisterCommand = async (e: FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    try {
-      setStatus("Loading...");
-      await axios.get(
-        `/api/discord-bot/register-commands?REGISTER_COMMANDS_KEY=${registerCommandsKey}`
-      );
-      setStatus("Commands Registered!!!");
-    } catch (error) {
-      setStatus("Something wen't wrong ");
-    }
-  };
-
-  return (
-    <>
-      <Head>
-        <title>Next Js Discord Bot Boilerplate | Mark Matthew Vergara</title>
-        <meta name="description" content="Generated by create next app" />
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-      <main className={`${inter.className}`}>
-        <h1>Nextjs Discord Bot Boilerplate</h1>
-        <a href="https://github.com/mmvergara/">
-          Template By: Mark Matthew Vergara
-        </a>
-        <a
-          href="https://github.com/mmvergara/mmv-nextjs-discord-bot-template"
-          target="_blank"
-        >
-          Github Repository
-        </a>
-
-        <form onSubmit={handleRegisterCommand}>
-          <input
-            type="text"
-            placeholder="Register Commands Key"
-            value={registerCommandsKey}
-            onChange={(e) => setRegisterCommandsKey(e.target.value)}
-          />
-          <button type="submit">Register Commands</button>
-        </form>
-        <a
-          className="hoverLinks"
-          href={`https://discord.com/api/oauth2/authorize?client_id=${process.env.NEXT_PUBLIC_CLIENT_APPLICATION_ID}&permissions=277025445888&scope=bot%20applications.commands`}
-          target="_blank"
-        >
-          Invite Discord Bot
-        </a>
-
-        <p style={{ paddingTop: 30 }}>{status}</p>
-      </main>
-    </>
-  );
-}
